requirements.txt
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
# Editable install with no version control (epsilon==0.1.0)
-e e:\edsec\epsilon
h11==0.16.0
idna==3.10
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mysql-connector-python==9.4.0
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.1.0
selenium==4.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
